[Nityanand]: Find Max among 3 Float Logic added

+++ MaximumNumber/MaximumNumberCheck.cs
@@ -3,7 +3,7 @@ namespace MaximumNumber
     public class MaximumNumberCheck
     {
         
-        public int MaxNumCheck(int fNum, int secNum, int thirNum)
+        public int MaxNumCheck(float fNum, float secNum, float thirNum)
         {
             if (fNum.CompareTo(secNum)>0 && fNum.ComapareTo(thirNum)>0)
             {
